에러 전체 수정

+++ SONIC_JK/jk_Time.cpp
@@ -23,8 +23,8 @@ namespace jk
 	{
 		QueryPerformanceCounter(&mCurFrequency);
 
-		double differnceFrequency = mCurFrequency.QuadPart - mPrevFrequency.QuadPart;
-
+		//double differnceFrequency = mCurFrequency.QuadPart - mPrevFrequency.QuadPart;
+		double differnceFrequency = static_cast<double>(mCurFrequency.QuadPart - mPrevFrequency.QuadPart);
